ip2gid: Move some macros and data structures to ip2gid.h
Move some macros and data structures to a spearate file. This is useful
for following patches.
